vis/requirements.txt
websockets
PyYaml
flask
